% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{sample_types}
\alias{sample_types}
\title{Helper function for sampling types of offspring upon division}
\usage{
sample_types(parent_type, Xi_as, Xi_s)
}
\arguments{
\item{parent_type}{ID of parent type}

\item{Xi_as}{matrix of asymmetric type transition probabilities}

\item{Xi_s}{matrix of symmetric type transition probabilities}
}
\description{
Helper function for sampling types of offspring upon division
}
