% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adb_ntaxa.R
\name{sim_adb_ntaxa_complete_fast}
\alias{sim_adb_ntaxa_complete_fast}
\title{Simulator of the complete Age-Dependent Branching Process (up to a fixed number of living particles)}
\usage{
sim_adb_ntaxa_complete_fast(
  ntaxa,
  a,
  b,
  d,
  origin_type = 0,
  Xi_as = matrix(0),
  Xi_s = matrix(1)
)
}
\arguments{
\item{ntaxa}{number of living particles at which the process is stopped (at least 2)}

\item{a}{vector of scale parameters per type}

\item{b}{vector of shape parameters per type}

\item{d}{vector of death probabilities per type}

\item{origin_type}{one of 0,...,n-1 where n is the number of types}

\item{Xi_as}{matrix of asymmetric type transition probabilities}

\item{Xi_s}{matrix of symmetric type transition probabilities}
}
\value{
a treedata object including dead particles, with \verb{@data} columns
\code{node}, \code{status} (0 = dead, 1 = alive, 2 = divided) and \code{type}, or
\code{NULL} (with a message) if too many particles died
}
\description{
Simulator of the complete Age-Dependent Branching Process (up to a fixed number of living particles)
}
