% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adb_ntaxa_gsa.R
\name{sim_adb_ntaxa_complete_gsa}
\alias{sim_adb_ntaxa_complete_gsa}
\title{Simulator of complete tree under ADB process with GSA}
\usage{
sim_adb_ntaxa_complete_gsa(
  ntaxa,
  scale = NULL,
  shape,
  death_prob,
  m = 2 * ntaxa,
  nsim = 10,
  max_attempts = 100,
  mean_lifetime = NULL
)
}
\arguments{
\item{ntaxa}{number of sampled particles (at least 2)}

\item{scale}{scale parameter (alternatively, give \code{mean_lifetime})}

\item{shape}{shape parameter}

\item{death_prob}{death probability}

\item{m}{number of standing taxa that will exist on the first generated trees, then be cut; m >> ntaxa}

\item{nsim}{number of trees to be generated, cut, and sampled from}

\item{max_attempts}{maximum number of simulation attempts per tree before
giving up (attempts fail when all particles die)}

\item{mean_lifetime}{mean lifetime (= scale * shape), alternative to \code{scale}}
}
\value{
a phylo object with \code{ntaxa} extant tips, including extinct lineages
}
\description{
Simulator of complete tree under ADB process with GSA
}
