% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adb_ntaxa.R
\name{sim_adb_ntaxa_samp}
\alias{sim_adb_ntaxa_samp}
\title{Simulator of a phylogeny from an Age-Dependent Branching process for a fixed number of sampled particles}
\usage{
sim_adb_ntaxa_samp(
  ntaxa,
  scale = NULL,
  shape,
  death_prob = 0,
  sampling_prob = 1,
  origin_type = 0,
  asym_trans_prob = matrix(0),
  sym_trans_prob = matrix(1),
  collapse = TRUE,
  mean_lifetime = NULL
)
}
\arguments{
\item{ntaxa}{number of sampled particles - tips in the phylogeny (at least 2)}

\item{scale}{vector of scale parameters per type (alternatively, give \code{mean_lifetime})}

\item{shape}{vector of shape parameters per type}

\item{death_prob}{vector of death probabilities per type}

\item{sampling_prob}{sampling probability}

\item{origin_type}{one of 0,...,n-1 where n is the number of types}

\item{asym_trans_prob}{matrix of asymmetric type transition probabilities}

\item{sym_trans_prob}{matrix of symmetric type transition probabilities}

\item{collapse}{if TRUE, nodes with a single descendant after pruning are
removed (node-typed tree), otherwise they are kept (branch-typed tree)}

\item{mean_lifetime}{vector of mean lifetimes per type (= scale * shape),
alternative to \code{scale}}
}
\value{
a treedata object with \code{ntaxa} tips, or \code{NULL} (with a message)
if too many particles died
}
\description{
Simulator of a phylogeny from an Age-Dependent Branching process for a fixed number of sampled particles
}
