% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adb_origin.R
\name{sim_adb_origin_complete_fast}
\alias{sim_adb_origin_complete_fast}
\title{R wrapper for a faster C++ implementation of complete Age-Dependent Branching Process
for a fixed time interval (since origin)}
\usage{
sim_adb_origin_complete_fast(
  origin_time,
  scale = NULL,
  shape,
  death_prob,
  origin_type = 0,
  asym_trans_prob = matrix(0),
  sym_trans_prob = matrix(1),
  min_taxa = 2,
  mean_lifetime = NULL
)
}
\arguments{
\item{origin_time}{time of birth of the initial particle}

\item{scale}{vector of scale parameters per type (alternatively, give \code{mean_lifetime})}

\item{shape}{vector of shape parameters per type}

\item{death_prob}{vector of death probabilities per type}

\item{origin_type}{one of 0,...,n-1 where n is the number of types}

\item{asym_trans_prob}{matrix of asymmetric type transition probabilities}

\item{sym_trans_prob}{matrix of symmetric type transition probabilities}

\item{min_taxa}{minimum number of tips in the tree}

\item{mean_lifetime}{vector of mean lifetimes per type (= scale * shape),
alternative to \code{scale}}
}
\value{
a treedata object including dead particles, with \verb{@data} columns
\code{node}, \code{status} (0 = dead, 1 = alive, 2 = divided) and \code{type}, or
\code{NULL} (with a message) if fewer than \code{min_taxa} particles are alive
}
\description{
R wrapper for a faster C++ implementation of complete Age-Dependent Branching Process
for a fixed time interval (since origin)
}
