% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{prune_tree}
\alias{prune_tree}
\title{Prune tree}
\usage{
prune_tree(obj, rho = NA, ntips = NA, min_tips = 2, collapse = TRUE)
}
\arguments{
\item{obj}{treedata object}

\item{rho}{sampling probability (at present only)}

\item{ntips}{number of sampled tips}

\item{min_tips}{minimum number of tips in the pruned tree}

\item{collapse}{if TRUE, stores the standard structure (node-typed tree), otherwise keeps nodes with one descendant (branch-typed tree)}
}
\description{
Prune tree
}
