% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adb_ntaxa.R
\name{sim_adb_ntaxa_complete_fast}
\alias{sim_adb_ntaxa_complete_fast}
\title{Simulator of the complete Age-Dependent Branching Process (up to a fixed number of living particles)}
\usage{
sim_adb_ntaxa_complete_fast(
  ntaxa,
  scale = NULL,
  shape,
  death_prob,
  origin_type = 0,
  asym_trans_prob = matrix(0),
  sym_trans_prob = matrix(1),
  mean_lifetime = NULL
)
}
\arguments{
\item{ntaxa}{number of living particles at which the process is stopped (at least 2)}

\item{scale}{vector of scale parameters per type (alternatively, give \code{mean_lifetime})}

\item{shape}{vector of shape parameters per type}

\item{death_prob}{vector of death probabilities per type}

\item{origin_type}{one of 0,...,n-1 where n is the number of types}

\item{asym_trans_prob}{matrix of asymmetric type transition probabilities}

\item{sym_trans_prob}{matrix of symmetric type transition probabilities}

\item{mean_lifetime}{vector of mean lifetimes per type (= scale * shape),
alternative to \code{scale}}
}
\value{
a treedata object including dead particles, with \verb{@data} columns
\code{node}, \code{status} (0 = dead, 1 = alive, 2 = divided) and \code{type}, or
\code{NULL} (with a message) if too many particles died
}
\description{
Simulator of the complete Age-Dependent Branching Process (up to a fixed number of living particles)
}
