% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adb_ntaxa_gsa.R
\name{sim_adb_ntaxa_complete_gsa}
\alias{sim_adb_ntaxa_complete_gsa}
\title{Simulator of complete tree under ADB process with GSA}
\usage{
sim_adb_ntaxa_complete_gsa(ntaxa, a, b, d, m = 2 * ntaxa, nsim = 10)
}
\arguments{
\item{ntaxa}{number of sampled particles (at least 2)}

\item{a}{scale parameter}

\item{b}{shape parameter}

\item{d}{death probability}

\item{m}{number of standing taxa that will exist on the first generated trees, then be cut; m >> ntaxa}

\item{nsim}{number of trees to be generated, cut, and sampled from}
}
\description{
Simulator of complete tree under ADB process with GSA
}
