% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adb_ntaxa.R
\name{sim_adb_ntaxa_samp}
\alias{sim_adb_ntaxa_samp}
\title{Simulator of a phylogeny from an Age-Dependent Branching process for a fixed number of sampled particles}
\usage{
sim_adb_ntaxa_samp(
  ntaxa,
  a,
  b,
  d = 0,
  rho = 1,
  origin_type = 0,
  Xi_as = matrix(0),
  Xi_s = matrix(1),
  collapse = TRUE
)
}
\arguments{
\item{ntaxa}{number of sampled particles - tips in the phylogeny (at least 2)}

\item{a}{vector of scale parameters per type}

\item{b}{vector of shape parameters per type}

\item{d}{vector of death probabilities per type}

\item{rho}{sampling probability}

\item{origin_type}{one of 0,...,n-1 where n is the number of types}

\item{Xi_as}{matrix of asymmetric type transition probabilities}

\item{Xi_s}{matrix of symmetric type transition probabilities}

\item{collapse}{if TRUE, nodes with a single descendant after pruning are
removed (node-typed tree), otherwise they are kept (branch-typed tree)}
}
\value{
a treedata object with \code{ntaxa} tips, or \code{NULL} (with a message)
if too many particles died
}
\description{
Simulator of a phylogeny from an Age-Dependent Branching process for a fixed number of sampled particles
}
