% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adb_ntaxa_gsa.R
\name{sim_adb_ntaxa_samp_gsa}
\alias{sim_adb_ntaxa_samp_gsa}
\title{Simulator of a phylogeny under ADB process with GSA}
\usage{
sim_adb_ntaxa_samp_gsa(
  ntaxa,
  a,
  b,
  d,
  rho,
  m = floor(2 * ntaxa/rho),
  nsim = 10,
  max_attempts = 100
)
}
\arguments{
\item{ntaxa}{number of sampled particles (at least 2)}

\item{a}{scale parameter}

\item{b}{shape parameter}

\item{d}{death probability}

\item{rho}{sampling probability}

\item{m}{number of standing taxa that will exist on the first generated trees (before sampling), then be cut; m >> ntaxa/rho}

\item{nsim}{number of trees to be generated, cut, and sampled from}

\item{max_attempts}{maximum number of simulation attempts per tree before
giving up (attempts fail when all particles die)}
}
\value{
a phylo object with \code{ntaxa} tips
}
\description{
Wrapper for general sampling approach (GSA) for tree simulation with n taxa
Compared to the simple sampling approach (SSA), this approach considers subsequent time periods with n particles alive
Note: so far, this approach can only be applied to single-type simulations
These functions return an object of class phylo
}
