% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_adb_origin.R
\name{sim_adb_origin_complete_fast}
\alias{sim_adb_origin_complete_fast}
\title{R wrapper for a faster C++ implementation of complete Age-Dependent Branching Process
for a fixed time interval (since origin)}
\usage{
sim_adb_origin_complete_fast(
  origin_time,
  a,
  b,
  d,
  origin_type = 0,
  Xi_as = matrix(0),
  Xi_s = matrix(1),
  min_tips = 2
)
}
\arguments{
\item{origin_time}{time of birth of the initial particle}

\item{a}{vector of scale parameters per type}

\item{b}{vector of shape parameters per type}

\item{d}{vector of death probabilities per type}

\item{origin_type}{one of 0,...,n-1 where n is the number of types}

\item{Xi_as}{matrix of asymmetric type transition probabilities}

\item{Xi_s}{matrix of symmetric type transition probabilities}

\item{min_tips}{minimum number of tips in the tree}
}
\description{
R wrapper for a faster C++ implementation of complete Age-Dependent Branching Process
for a fixed time interval (since origin)
}
